+++ packages/icons/BarChart.svelte
@@ -1,0 +1,21 @@
+<!-- This Source Code Form is subject to the terms of the Mozilla Public
+   - License, v. 2.0. If a copy of the MPL was not distributed with this
+   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
+
+<!-- This icon was contributed by the graph-paper team. -->
+<script>
+export let size = 24;
+</script>
+<svg width={size} height={size} viewBox="0 0 20 20" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" xml:space="preserve" xmlns:serif="http://www.serif.com/" style="fill-rule:evenodd;clip-rule:evenodd;stroke-linejoin:round;stroke-miterlimit:2;" fill=currentColor>
+    <g transform="matrix(1,0,0,1,-2,-2)">
+        <g transform="matrix(1.375,0,0,1.11111,-0.75,-2.44444)">
+            <path d="M2,4L6,4L6,22L2,22L2,4"/>
+        </g>
+        <g transform="matrix(1.375,0,0,0.5,6.25,11)">
+            <path d="M2,4L6,4L6,22L2,22L2,4"/>
+        </g>
+        <g transform="matrix(1.375,0,0,0.833333,13.75,3.66667)">
+            <path d="M2,4L6,4L6,22L2,22L2,4"/>
+        </g>
+    </g>
+</svg>