+++ packages/icons/LineChart.svelte
@@ -1,0 +1,16 @@
+<!-- This Source Code Form is subject to the terms of the Mozilla Public
+   - License, v. 2.0. If a copy of the MPL was not distributed with this
+   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
+
+<!-- This icon was contributed by the graph-paper team. -->
+
+<script>
+export let size = 24;
+</script>
+
+
+<svg width={size} height={size} viewBox="0 0 24 24" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" xml:space="preserve" xmlns:serif="http://www.serif.com/" style="fill-rule:evenodd;clip-rule:evenodd;stroke-linejoin:round;stroke-miterlimit:2;">
+    <g transform="matrix(0.696392,-0.696392,-0.733892,-0.733892,10.9822,28.1995)">
+        <path fill=currentColor d="M2,14L2,12L10,12L10,7.012C10,6.744 10.112,6.486 10.312,6.297C10.513,6.107 10.784,6 11.067,6C14.005,6 22,6 22,6L22,8L12,8L12,13.138C12,13.614 11.593,14 11.091,14C8.626,14 2,14 2,14Z"/>
+    </g>
+</svg>