+++ packages/icons/base.js
@@ -1,0 +1,53 @@
+import Close from "./Close.svelte";
+import ArrowheadDown from "./ArrowheadDown.svelte";
+import ArrowheadUp from "./ArrowheadUp.svelte";
+import ArrowheadLeft from "./ArrowheadLeft.svelte";
+import ArrowheadRight from "./ArrowheadRight.svelte";
+import Forward from "./Forward.svelte";
+import Help from "./Help.svelte";
+import Info from "./Info.svelte";
+import Information from "./Information.svelte";
+import Menu from "./Menu.svelte";
+import Search from "./Search.svelte";
+import Warning from "./Warning.svelte";
+
+import LineChart from "./LineChart.svelte";
+import BarChart from "./BarChart.svelte";
+import Table from "./Table.svelte";
+
+const icons = {
+  Close,
+  ArrowheadUp,
+  ArrowheadDown,
+  ArrowheadLeft,
+  ArrowheadRight,
+  Forward,
+  Help,
+  Info,
+  Information,
+  Menu,
+  Search,
+  Warning,
+  BarChart,
+  Table,
+  LineChart,
+};
+export default icons;
+
+export {
+  Close,
+  ArrowheadUp,
+  ArrowheadDown,
+  ArrowheadLeft,
+  ArrowheadRight,
+  Forward,
+  Help,
+  Info,
+  Information,
+  Menu,
+  Search,
+  Warning,
+  BarChart,
+  Table,
+  LineChart,
+};