+++ packages/icons/stories/AllIcons.svelte
@@ -1,0 +1,33 @@
+<script>
+import icons from '..';
+
+export let size = 24;
+export let color = '#000';
+</script>
+
+<style>
+.icon-set {
+  display: grid;
+  grid-template-columns: max-content max-content;
+  align-items: center;
+  grid-column-gap: var(--space-2x);
+  grid-row-gap: var(--space-base);
+}
+
+.icon--component {
+  display: grid;
+  align-items: center;
+  justify-items: center;
+}
+</style>
+
+<div class="icon-set">
+  {#each Object.entries(icons) as [name, icon], i}
+    <div style='color: {color};' class="icon--component">
+      <svelte:component this={icon} {size} />
+    </div>
+  <div class="icon--name">
+      {name}
+    </div>
+  {/each}
+</div>